Added overwrite course discussion specs function to course utilities and renamed that
