+++ src/generalUtilities/CourseUtilities.ts
@@ -1,5 +1,7 @@
-import { StudentScoreData } from "../generalModels/DiscussionScoring";
+import { DiscussionSpecs, StudentScoreData } from "../generalModels/DiscussionScoring";
 import { Course, courseModel } from "../generalModels/Course";
+
+const DATABASE_ERROR_MESSAGE = "Database error. Please message admin";
 
 export async function getCourseByName(courseName: string) {
     let course: Course | null = null;
@@ -17,4 +19,19 @@ export async function getCourseByName(courseName: string) {
     })
 
     return course !== null ? course : undefined;
+}
+
+export async function overwriteCourseDiscussionSpecs(courseName: string, discussionSpecs: DiscussionSpecs): Promise<string> {
+    
+    try {
+        await courseModel.findOneAndUpdate( 
+            {name: courseName}, 
+            {"discussionSpecs": discussionSpecs}
+        )
+    }
+    catch(error: any) {
+        console.error(error);
+        return DATABASE_ERROR_MESSAGE;
+    }
+    return "";
 }
