+++ src/pieces/discussion/menu/staff/ViewStudentsMenu.ts
@@ -2,7 +2,7 @@ import { APIEmbedField, ButtonStyle, Guild, GuildMember, InteractionUpdateOption
 import { Course } from "../../../../generalModels/Course";
 import { ScorePeriod, StudentScoreData } from "../../../../generalModels/DiscussionScoring";
 import { sendDismissableReply } from "../../../../generalUtilities/DismissableMessage";
-import { getCourseByName } from "../../../../generalUtilities/getCourseByName";
+import { getCourseByName } from "../../../../generalUtilities/CourseUtilities";
 import { GUILDS } from "../../../../secret";
 import { addScorePeriods } from "../../scoring/scoreFunctions";
 import { ComponentBehavior } from "../../../menu/BaseMenu";
